# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "tdPSA API Documentation",
        "description": "API documentation for tdPSA Clients module",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "tdPSA API Server"
        }
    ],
    "paths": {
...
        "/api/v1/sales/opportunities": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get sales opportunities",
                "operationId": "getSalesOpportunityList",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "owner_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Missing sales.read scope"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create sales opportunity",
                "operationId": "createSalesOpportunity",
                "responses": {
                    "201": {
                        "description": "Opportunity created"
                    },
                    "403": {
                        "description": "Missing sales.create scope"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sales/opportunities/{opportunity}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get sales opportunity",
                "operationId": "getSalesOpportunityByKey",
                "parameters": [
                    {
                        "name": "opportunity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Missing sales.read scope"
                    },
                    "404": {
                        "description": "Opportunity not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Sales"
                ],
                "summary": "Update sales opportunity",
                "operationId": "updateSalesOpportunity",
                "parameters": [
                    {
                        "name": "opportunity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opportunity updated"
                    },
                    "403": {
                        "description": "Missing sales.update scope"
                    },
                    "404": {
                        "description": "Opportunity not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sales/opportunities/{opportunity}/activities": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create sales activity",
                "operationId": "createSalesActivity",
                "parameters": [
                    {
                        "name": "opportunity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Activity created"
                    },
                    "403": {
                        "description": "Missing sales.update scope"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/sales/opportunities/{opportunity}/read": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Mark sales opportunity activities read",
                "operationId": "markSalesOpportunityRead",
                "parameters": [
                    {
                        "name": "opportunity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opportunity marked read"
                    },
                    "403": {
                        "description": "Missing sales.update scope"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}